Basic Authentication added

+++ src/main/java/com/drag/ftptowebpictures/repository/FileRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import com.drag.ftptowebpictures.model.File;
 
 @RepositoryRestResource
-public interface FileRepository extends JpaRepository<File, Long> {
+public interface FileRepository extends JpaRepository<File, Integer> {
 
 }
